+++ apps/extension/src/@talisman/components/SuspenseTracker.tsx
@@ -15,7 +15,5 @@ export const SuspenseTracker: FC<{ name: string }> = ({ name }) => {
     }
   }, [name])
 
-  console.log("[Suspense] " + name)
-
   return null
 }
