+++ apps/extension/src/ui/hooks/useFirstAccountColors.ts
@@ -1,18 +1,7 @@
-import { useTalismanOrb } from "@talismn/orb"
 import { accountsState } from "@ui/atoms"
-import { useMemo } from "react"
 import { useRecoilValue } from "recoil"
 
-const TALISMAN_COLORS = ["#fd4848", "#d5ff5c"] as const
-
-export const useAccountColors = (address?: string) => {
-  const { bgColor1, bgColor2 } = useTalismanOrb(address ?? "")
-
-  return useMemo(
-    () => (address ? [bgColor1, bgColor2] : TALISMAN_COLORS) as [string, string],
-    [address, bgColor1, bgColor2]
-  )
-}
+import { useAccountColors } from "./useAccountColors"
 
 export const useFirstAccountColors = () => {
   // pick first account and apply it's colors to background
